refactor: functions for geometry render
